Routes protection via group user permissions (partial implementation).

+++ packages/backend/src/panel/panel.controller.ts
@@ -3,63 +3,59 @@ import { PanelService } from "./panel.service";
 import { ApiBearerAuth, ApiCreatedResponse, ApiUseTags } from "@nestjs/swagger";
 import { AuthGuard } from "@nestjs/passport";
 import { CreatePanelDto, DuplicatePanelDto, PanelDto, UpdatePanelDto } from "@airlab/shared/lib/panel/dto";
-import { JwtPayloadDto } from "@airlab/shared/lib/auth/dto";
 import { GroupUserService } from "../groupUser/groupUser.service";
 
 @Controller()
+@UseGuards(AuthGuard("jwt"))
 @ApiUseTags("panels")
 @ApiBearerAuth()
-@UseGuards(AuthGuard("jwt"))
 export class PanelController {
   constructor(private readonly panelService: PanelService, private readonly groupUserService: GroupUserService) {}
-
-  @Get("panels")
-  @ApiCreatedResponse({ description: "Find all entities.", type: PanelDto, isArray: true })
-  findAll() {
-    return this.panelService.findAll();
-  }
-
-  @Get("panels/:id")
-  @ApiCreatedResponse({ description: "Find entity by Id.", type: PanelDto })
-  findById(@Param("id") id: number) {
-    return this.panelService.findById(id);
-  }
 
   @Post("panels")
   @ApiCreatedResponse({ description: "Create entity.", type: PanelDto })
   async create(@Request() req, @Body() params: CreatePanelDto) {
-    const user: JwtPayloadDto = req.user;
-    const groupUser = await this.groupUserService.findByUserIdAndGroupId(user.userId, params.groupId);
+    const groupUser = await this.groupUserService.checkGroupUserPermissions(req.user.userId, params.groupId);
     return this.panelService.create({ ...params, createdBy: groupUser.id });
+  }
+
+  @Get("panels/:id")
+  @ApiCreatedResponse({ description: "Find entity by Id.", type: PanelDto })
+  async findById(@Request() req, @Param("id") id: number) {
+    const item = await this.panelService.findById(id);
+    await this.groupUserService.checkGroupUserPermissions(req.user.userId, item.groupId);
+    return item;
   }
 
   @Patch("panels/:id")
   @ApiCreatedResponse({ description: "Updated entity.", type: PanelDto })
-  async update(@Param("id") id: number, @Body() params: UpdatePanelDto) {
+  async update(@Request() req, @Param("id") id: number, @Body() params: UpdatePanelDto) {
+    const item = await this.panelService.findById(id);
+    await this.groupUserService.checkGroupUserPermissions(req.user.userId, item.groupId);
     return this.panelService.update(id, params);
   }
 
   @Put("panels/:id")
   @ApiCreatedResponse({ description: "Duplicate entity.", type: PanelDto })
   async duplicate(@Request() req, @Param("id") id: number, @Body() params: DuplicatePanelDto) {
-    const user: JwtPayloadDto = req.user;
-    const groupUser = await this.groupUserService.findByUserIdAndGroupId(user.userId, params.groupId);
+    const groupUser = await this.groupUserService.checkGroupUserPermissions(req.user.userId, params.groupId);
     return this.panelService.duplicate(id, { ...params, createdBy: groupUser.id });
   }
 
   @Delete("panels/:id")
   @ApiCreatedResponse({ description: "Delete entity by Id.", type: Number })
-  deleteById(@Param("id") id: number) {
+  async deleteById(@Request() req, @Param("id") id: number) {
+    const item = await this.panelService.findById(id);
+    await this.groupUserService.checkGroupUserPermissions(req.user.userId, item.groupId);
     return this.panelService.deleteById(id);
   }
 
-  @Get("group/:groupId/panels")
+  @Get("groups/:groupId/panels")
   @ApiCreatedResponse({ description: "Find all panels for the group.", type: PanelDto, isArray: true })
   async getAllPanelsForGroup(@Request() req, @Param("groupId") groupId: number) {
-    const user: JwtPayloadDto = req.user;
-    const groupUser = await this.groupUserService.findByUserIdAndGroupId(user.userId, groupId);
+    const groupUser = await this.groupUserService.checkGroupUserPermissions(req.user.userId, groupId);
     return groupUser.canPanels
-      ? this.panelService.getAllPanelsForGroup(groupId)
-      : this.panelService.getAllPanelsForGroupUser(groupUser.id);
+      ? this.panelService.getGroupPanels(groupId)
+      : this.panelService.getPersonalGroupPanels(groupId, groupUser.id);
   }
 }
